Fixed effect meta-analysis for correlated test statistics

Fixed effect meta-analysis for correlated test statistics using the Lin-Sullivan method.

LS(beta, stders, cor = diag(1, length(beta)))

Arguments

beta
regression coefficients from each analysis

stders
standard errors corresponding to betas

cor
correlation matrix between of test statistics. Default considers uncorrelated test statistics

Value

Return values:

beta: effect size

se: effect size standard error

p: p-value

Details

Perform fixed effect meta-analysis for correlated test statistics using method of Lin and Sullivan (2009). By default, correlation is set to identity matrix to for independent test statistics.

This method requires the correlation matrix to be symmatric positive definite (SPD). If this condition is not satisfied, results will be NA. If the matrix is not SPD, there is likely an issue with how it was generated.

However, evaluating the correlation between observations that are not pairwise complete can give correlation matricies that are not SPD. In this case, consider running Matrix::nearPD( x, corr=TRUE) to produce the nearest SPD matrix to the input.

References

Lin D, Sullivan PF (2009). “Meta-analysis of genome-wide association studies with overlapping subjects.” The American Journal of Human Genetics, 85(6), 862--872.

Examples

# Generate effects
library(mvtnorm)
library(clusterGeneration )

n = 4
Sigma = cov2cor(genPositiveDefMat(n)$Sigma)
beta = t(rmvnorm(1, rep(0, n), Sigma))
stders = rep(.1, n)

# Run fixed effects meta-analysis, 
# assume identity correlation  
LS( beta, stders)

#>          beta   se         p
#> 1 -0.04733689 0.05 0.3437724

# Run random effects meta-analysis,
# assume identity correlation  
RE2C( beta, stders)

#>       stat1    stat2        RE2Cp RE2Cp.twoStep      QE           QEp      Isq
#> 1 0.8963124 687.4698 9.45509e-151            NA 712.198 4.753493e-154 99.57877

# Run fixed effects meta-analysis, 
# account for correlation 
LS( beta, stders, Sigma)

#>          beta         se          p
#> 1 -0.07322542 0.03482681 0.03550434

# Run random effects meta-analysis,
# account for correlation 
RE2C( beta, stders, Sigma)

#>      stat1    stat2         RE2Cp RE2Cp.twoStep QE QEp Isq
#> 1 4.420754 478.7468 3.876244e-106            NA NA  NA  NA