Random effect meta-analysis for correlated test statistics

Random effect meta-analysis for correlated test statistics using RE2C

Usage

RE2C(beta, stders, cor = diag(1, length(beta)), twoStep = FALSE)

Arguments

beta

regression coefficients from each analysis

stders

standard errors corresponding to betas

cor

correlation matrix between of test statistics. Default considers uncorrelated test statistics

twoStep

Apply two step version of RE2C that is designed to be applied only after the fixed effect model.

Value

Return values:

stat1: statistic testing effect mean

stat2: statistic testing effect heterogeneity

RE2Cp: RE2 p-value accounting for correlelation between tests

RE2Cp.twoStep: two step RE2C test after fixed effect test. Only evaluated if twoStep==TRUE

QE: test statistic for the test of (residual) heterogeneity

QEp: p-value for the test of (residual) heterogeneity

Isq: I^2 statistic

QE, QEp and ISq are only evaluted if correlation is diagonal

Details

Perform random effect meta-analysis for correlated test statistics using RE2 method of Han and Eskin (2011), or RE2 for correlated test statistics from Han, et al., (2016). Also uses RE2C method of Lee, Eskin and Han (2017) to further test for heterogenity in effect size. By default, correlation is set to identity matrix to for independent test statistics.

This method requires the correlation matrix to be symmatric positive definite (SPD). If this condition is not satisfied, results will be NA. If the matrix is not SPD, there is likely an issue with how it was generated.

However, evaluating the correlation between observations that are not pairwise complete can give correlation matricies that are not SPD. In this case, consider running Matrix::nearPD( x, corr=TRUE) to produce the nearest SPD matrix to the input.

References

Lee CH, Eskin E, Han B (2017). “Increasing the power of meta-analysis of genome-wide association studies to detect heterogeneous effects.” Bioinformatics, 33(14), i379--i388. https://doi.org/10.1093/bioinformatics/btx242. Han B, Duong D, Sul JH, de Bakker PI, Eskin E, Raychaudhuri S (2016). “A general framework for meta-analyzing dependent studies with overlapping subjects in association mapping.” Human Molecular Genetics, 25(9), 1857--1866. https://doi.org/10.1093/hmg/ddw049. Han B, Eskin E (2011). “Random-effects model aimed at discovering associations in meta-analysis of genome-wide association studies.” The American Journal of Human Genetics, 88(5), 586--598. https://doi.org/10.1016/j.ajhg.2011.04.014.

Examples

# Generate effects
library(mvtnorm)
library(clusterGeneration )

set.seed(1)
n = 4
Sigma = cov2cor(genPositiveDefMat(n)$Sigma)
beta = t(rmvnorm(1, rep(0, n), Sigma))
stders = rep(.1, n)

# Run fixed effects meta-analysis, 
# assume identity correlation  
LS( beta, stders)
#>         beta   se          p
#> 1 0.09877893 0.05 0.04820252

# Run random effects meta-analysis,
# assume identity correlation  
RE2C( beta, stders)
#>      stat1    stat2        RE2Cp RE2Cp.twoStep       QE          QEp      Isq
#> 1 3.902911 124.5332 4.031684e-29            NA 142.8348 9.250987e-31 97.89967

# Run fixed effects meta-analysis, 
# account for correlation 
LS( beta, stders, Sigma)
#>        beta         se            p
#> 1 0.1806549 0.03973155 5.444484e-06

# Run random effects meta-analysis,
# account for correlation 
RE2C( beta, stders, Sigma)
#>      stat1    stat2       RE2Cp RE2Cp.twoStep QE QEp Isq
#> 1 20.67419 117.2758 3.77186e-31            NA NA  NA  NA